Split flow control package into multiple files
